Weather-A2/requirements.txt
pandas
protobuf
python-dotenv
requests
streamlit-extras
streamlit_shadcn_ui
streamlit_autorefresh
plotly
os
sys
streamlit
datetime
time
yagmail
google-generativeai
schedule
firebase_admin

